requirements.txt
amqp==2.4.2
asgiref==3.3.1
attrs==19.1.0
autobahn==21.2.1
Automat==0.8.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
cached-property==1.5.1
cachetools==3.1.0
celery==4.3.0
certifi==2019.3.9
cffi==1.14.0
channels==3.0.4
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==2.9.2
daphne==3.0.2
defusedxml==0.6.0
Django==3.0.5
django-admin-multiple-choice-list-filter==0.1.1
django-autotranslate==1.1.1
django-bootstrap-v5==1.0.11
django-braces==1.14.0
django-ckeditor==5.6.1
django-cors-headers==3.8.0
django-dbbackup==3.2.0
django-dbbackup-ui==0.2.1
django-filter==2.1.0
django-fsm==2.7.0
django-js-asset==1.2.2
django-modeltranslation==0.15
django-mptt==0.11.0
django-multiurl==1.4.0
django-nested-admin==3.2.2
django-oauth-toolkit==1.3.2
django-polymorphic==3.0.0
django-recaptcha3==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-simple-history==2.7.0
django-solo==1.1.3
django-utils-six==2.0
django-uuidfield==0.5.0
django-uuslug==1.2.0
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
docopt==0.6.2
docxtpl==0.11.3
drf-yasg==1.17.1
easy-thumbnails==2.7
entrypoints==0.3
et-xmlfile==1.1.0
fcm-django==0.3.4
ffmpeg-python==0.2.0
flake8==3.7.9
future==0.18.2
futures==3.1.1
garpix-admin @ file:///home/polski-urody-back/vendor/garpix-admin-0.1.4.tar.gz
garpix-auth @ file:///home/polski-urody-back/vendor/garpix-auth-0.2.1.tar.gz
garpix-confirm @ file:///home/polski-urody-back/vendor/garpix-confirm-0.1.4.tar.gz
garpix-event @ file:///home/polski-urody-back/vendor/garpix-event-0.3.3.tar.gz
garpix-menu @ file:///home/polski-urody-back/vendor/garpix-menu-0.1.15.tar.gz
garpix-multiupload==1.0.0
garpix-notify @ file:///home/polski-urody-back/vendor/garpix-notify-0.3.30.tar.gz
garpix-page @ file:///home/polski-urody-back/vendor/garpix-page-0.6.0.tar.gz
garpix-pay @ file:///home/polski-urody-back/vendor/garpix-pay-0.2.1.tar.gz
garpix-qa @ file:///home/polski-urody-back/vendor/garpix-qa-0.1.3.tar.gz
garpix-reclaim @ file:///home/polski-urody-back/vendor/garpix-reclaim-0.1.5.tar.gz
garpix-subscribe @ file:///home/polski-urody-back/vendor/garpix-subscribe-0.1.2.tar.gz
garpix-utils @ file:///home/polski-urody-back/vendor/garpix-utils-0.5.4.tar.gz
google-api-core==1.9.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-translate==1.4.0
googleapis-common-protos==1.5.9
goslate==1.5.1
grpcio==1.28.1
gunicorn==19.9.0
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==21.3.0
inflection==0.4.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.0.1
kombu==4.5.0
lxml==4.5.0
Markdown==3.1
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==1.6.0
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==2.5.14
packaging==19.0
pdfkit==0.6.1
Pillow==8.4.0
pip-review==1.0
polib==1.1.0
protobuf==3.7.1
psycopg2-binary==2.9.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.20
pyfcm==1.4.7
pyflakes==2.1.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-docx==0.8.11
python-dotenv==0.10.1
python-monkey-business==1.0.0
python-slugify==3.0.2
python3-openid==3.1.0
pytils==0.3
pytz==2019.1
PyYAML==5.1
raven==6.10.0
redis==3.2.1
requests==2.21.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.3.1
sqlparse==0.3.0
text-unidecode==1.2
texttable==1.6.1
transliterate==1.10.2
Twisted==19.10.0
txaio==21.2.1
Unidecode==1.0.23
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
websocket-client==0.56.0
zipp==3.1.0
zope.interface==5.4.0
